Adds branch:main to install instructions
When testing this on a project, I encountered an issue trying to install this through bundler. Bundler appears to automatically look for a branch named 'master' on gems installed via git. This central branch for this repo is main and so bundler throws an error. This is a suggestion to update the README with the branch: 'main' so that others can more easily get started with this gem. :)

Additionally fixes line 427 to use backticks to enable code highlighting of `option_`