Introduce `:region` selector
While similar to the `:section` selector, the `:region` selector
_requires_ an `[aria-label]` or `[aria-labelledby]` attribute as a
locator, instead of an element's first `<h1>` through `<h6>` heading
element.

In accordance with the [ARIA Authoring Practices: Landmark Regions][]
section on the [Region][] landmark, add selector support for finding
`<section>` or `[role="region"]` elements.

A [region][] landmark is a perceivable section of the page containing
content that is sufficiently important for users to be able to navigate
to the section.

* A region landmark must have a label.
* If a page includes more than one region landmark, each should have a
  unique label (see [Step 3][] above).
* The region landmark can be used identify content that named landmarks
  do not appropriately describe.

[ARIA Authoring Practices: Landmark Regions]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/
[Region]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/#x4-3-7-region
[region]: https://w3c.github.io/aria/#region
[Step 3]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/#aria_lh_step3