Introduce `:contentinfo` selector
In accordance with the [ARIA Authoring Practices: Landmark Regions][]
section on the [Contentinfo][] landmark, add selector support for finding
`<header>` or `[role="contentinfo"]` elements that are direct descendants of
the `<body>` element.

A [contentinfo][] landmark is a way to identify common information at
the bottom of each page within a website, typically called the "footer"
of the page, including information such as copyrights and links to
privacy and accessibility statements.

* Each page may have one contentinfo landmark.
* The contentinfo landmark should be a top-level landmark.
* When a page contains nested document and/or application roles (e.g.
  typically through the use of iframe and frame elements), each document
or application role may have one contentinfo landmark.
* If a page includes more than one contentinfo landmark, each should
  have a unique label (see [Step 3][] above).

[ARIA Authoring Practices: Landmark Regions]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/
[Contentinfo]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/#x4-3-1-contentinfo
[contentinfo]: https://w3c.github.io/aria/#contentinfo
[Step 3]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/#aria_lh_step3