Introduce `:banner` selector (#44)
In accordance with the [ARIA Authoring Practices: Landmark Regions][]
section on the [Banner][] landmark, add selector support for finding
`<header>` or `[role="banner"]` elements that are direct descendants of
the `<body>` element.

A [banner][] landmark identifies site-oriented content at the beginning of
each page within a website. Site-oriented content typically includes
things such as the logo or identity of the site sponsor, and
site-specific search tool. A banner usually appears at the top of the
page and typically spans the full width.

* Each page may have one banner landmark.
* The banner landmark should be a top-level landmark.
* When a page contains nested document and/or application roles (e.g.
  typically through the use of iframe and frame elements), each document
  or application role may have one banner landmark.
* If a page includes more than one banner landmark, each should have a
  unique label (see [Step 3][] above).

[ARIA Authoring Practices: Landmark Regions]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/
[Banner]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/#x4-3-1-banner
[banner]: https://w3c.github.io/aria/#banner
[Step 3]: https://www.w3.org/WAI/ARIA/apg/practices/landmark-regions/#aria_lh_step3

Co-authored-by: Daniel Lewis <github@mrdaniellewis.co.uk>